requirements.txt
tensorflow>=2.5.0
scikit-learn
Flask
opencv-python
numpy
joblib
torch==1.9.0
torchvision==0.10.0
Pillow==8.2.0

